Final-DCT/python/requirements.txt
numpy
matplotlib
pillow
scipy
pytest
pandas
